rag_project/flask-login-app/requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
flask-wtf==1.1.1
wtforms==3.0.1